pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-node-claude-scaffold"
version = "0.1.0"
description = "A clean, forkable scaffold for Claude-powered Python applications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv",
    "pydantic>=2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit",
]

[project.scripts]
claude-agent = "src.agent.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.11"
exclude = ["tests/", "src/scripts/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["src/scripts/*"]
